coalesce_writer: Fix race condition
The internal bufio.Writer ('buf') is being accessed in multiple
different goroutines. This means that if we Write() to the writer
when a flush is happening, the buf can get into an inconsistent state
and return io.ErrShortWrite when attempting to flush.

For more information, check the source for bufio.Writer. The relevant
code there is writer.flush() updating writer.n and writer.Write() also
updating writer.n (as of rev 00a411da5f5a).