refactor(ioutils): remove goroutine in LimitWriter
Yet again, I think using a mutex makes for cleaner code and is less
complicated. The goroutine was added to make sure we only interact with
the buffer in a single goroutine at any given moment, but since the
synchronization needed is minimal a Mutex works just as well.

Close #28.