Remove everything and start over
This implementation was pretty much a spike and some of my first Go code. I've
read through it several times and attempted to refactor it to the way I want it,
but I finally decided that just starting over would be the best way to implement
all the things I want (better error handling, better testability, etc.).