Add a basic integration test
This is meant to test the "happy path" of a job, i.e. just push a
regular build onto the AMQP queue and then verify that the expected
outputs happened (hub and logs AMQP queues, and that the backend got
the right script).

Only some things are being verified right now, but this is only meant
to be a start.