job-queue: ack if we can't parse JSON
This is the behaviour of the Ruby worker and how we deal with invalid
JSON from hub. I kept it as a reject-and-requeue for a while as the
parse errors were related to invalid struct fields so the Ruby worker
could deal with the message instead, but after some testing I'm now
confident that these errors are gone.