backend: don't export implementations of Provider and Instance
The providers and instance implementations don't implement anything
that's useful outside of the package that is not already a part of the
Provider and Instance interfaces. The providers should be created using
NewProvider and the instances should be created through the provider, so
I see no reason to keep them exported.