log-writer: close the log writer in the run-script step
Previously, this was done in the backend, before the SSH session was
closed, which I suspect caused a race condition where sometimes the log
writer would be closed before all log data was flushed to the log
writer.

There also seems to have been a bug in the Docker backend causing the
log writer never to be closed.