step/generate-script: retry in-client instead of requeueing job
The retry logic is implemented using exponential backoff. The default
values for the backoff module allowed for a retry interval to be up to a
minute and the total duration to be 15 minutes, which I found a bit
excessive, so I reduced it to a 10 second interval and total duration of
a minute.