processor: close channel instead of send for graceful shutdown
The graceful channel isn't buffered, so sending to it will block until
the processor is shut down. Calling GracefulShutdown() again will block
forever. Closing the channel will return immediately.

This should fix the issue with graceful shutdowns, since the processor
pool waits for GracefulShutdown() to return before moving on to the next
processor.