Update gometalinter on Travis CI
Just running `go get` means we'll just ensure it is installed, but since it is
preinstalled, it won't get updated. Then, when we run `gometalinter -i -u` later
on, it will update `gometalinter`. Between the preinstalled version and the
current version, more linters were installed, so when trying to then run
`gometalinter`, we get an error due to some linters not being installed.