Fix systemd service file
EnvironmentFiles aren't "executed", and since we depend on that in order
to have certain variables be defined by using others, we can't use them.
Changed to a /bin/sh -c '' style command instead in ExecStart, similar
to the upstart file.

In addition, the binary name in Exec* needs to be a full path, so
/bin/mkdir instead of just mkdir, for example.