ssh: extract common SSH logic to a subpackage
The different backends were doing a lot of common SSH logic that is more
or less the same across all packages. This extracts it to a subpackage
that they all can use without having to re-do all of the logic. This
also brings the recent improvements to the SSH logic in the jupiterbrain
backend over to the other backends.