ssh: request a PTY that's 80 columns wide and 40 rows tall
The RequestPty function takes the size as height and width, not width
and height, so we were requesting the somewhat non-standard size of 40
columns wide and 80 rows tall. A terminal width of 80 is somewhat common
to emulate (although most terminals are probably even wider than this).

This should only affect commands that change their output dynamically
based on the dimensions on the terminal.