check tagsets for trailing commas before querying job-board
When .travis.yml contains trailing commas, those commas propagate all
the way through to to worker, which builds them into tags and passes
those tags on to job-board.

job-board does not like tags with commas and returns a 500, which then
causes a retry loop.

This patch adds a clearer error message for when that happens, displays
this error message to the user, and also makes sure we do not requeue.