log-writer: pass timeout on creation and start timer on first write
Previous to this, if you did a Write() before calling SetTimeout, the
Write would cause the timeout timer to be set to 0 seconds, which would
cause it to immediately time out (though, due to a race condition, this
only occasionally happened since the step after the first write (the
worker info write) is calling SetTimeout, so the Go runtime wouldn't
always switch to the timer before it got set again.

This change both resolves the race condition and also prevents it from
happening again by passing the timeout in when creating the log writer,
and then starting the timer on the first write.