backend/docker: change SHM default to 64 MiB
The actual default is defined later to be 64 * 1024 * 1024 bytes:

    shm := uint64(1024 * 1024 * 64)

If it's overriden with SHM, say by saying "SHM=64MB" (which previously
was indicated to be the default in the help text, it would go through
humanize.ParseBytes:

    if cfg.IsSet("SHM") {
        if parsedShm, err := humanize.ParseBytes(cfg.Get("SHM")); err == nil {
          shm = parsedShm
        }
    }

humanize.ParseBytes("64MB") returns 64 * 1000 * 1000, not
64 * 1024 * 1024, so if you specify SHM=64MB you get something other
than the default of "64MB". By changing the default in the docs text to
"64MiB", the docs default reflects the actual default, as you could pass
SHM=64MiB and get the same value as the default.

Note that Docker parses numbers differently, so if you pass --shm=64MB
on the Docker command line, you end up with 64 * 1024 * 1024 bytes.