Revert "build: cache vendor/ directory on Travis CI"
This caches the vendor/manifest file as well, which means that updated
dependencies won't be updated, since the cache overwrites what's in the
git repo.

Also, this directory only contains the downloaded files, so the speed
increase wasn't as significant as I first thought. The intention was to
cache the compiled files for the dependencies, not just the source code.

This reverts commit 876aaf9911c86c044cfac82d0a9a3df652a0ff07.