Allow multiple cancellation subscriptions for same job ID
Prior to this, you could have a job get requeued back to the same worker
immediately after an error, but since there are some cleanup steps
between the requeue and when the job actually "finishes" in worker, and
the "unsubscribe cancellation" step is done last, you could then have a
situation where a new job is redelivered to the same worker before the
old processor has unsubscribed.

This commit also uses the opportunity to refactor the cancellation logic
a bit and pull the "broadcast" part to a separate type instead of
embedding it in the old Canceller interface and requiring each
implementor to implement the semantics correctly.

This makes it easier to test the semantics of cancellation without
setting up a "Canceller", you just use the in-memory cancellation
broadcaster and send cancellation broadcasts to it. All the
AMQPCanceller type does now is listen to AMQP commands and forward
cancellations to the CancellationBroadcaster.

The FileCanceller was removed since it was a no-op and there's no longer
a need for something to implement the Canceller interface.

Fixes #288.