Close job channel when finished
The processor checks for this to be closed, and terminates when it has.
If we never close it, there's a possibility that this function returns
and the processor ends up just being stuck forever, which means the
worker could be in a "zombie" state where it's still running but no
longer accepting new jobs.

Most of the time, anything that would cause this loop to exit, such as
AMQP errors, would also cause context cancellations, which would cause
the processor to exit, but I don't think we've exhaustively checked
every option, so I think it's safer to just close the channel here.