introduce log-metrics flag, change log metrics to opt-in
When running worker without librato configured, e.g. locally or on enterprise,
we will print some metrics to the console once per minute. This output tends
to be quite spammy, and doesn't make for a very nice out-of-the-box experience.

It's possible to opt-out of that behaviour by specifying:

```
TRAVIS_WORKER_SILENCE_METRICS=true
```

This patch is intended to make the default experience more user friendly by
changing the behaviour to an opt-in `TRAVIS_WORKER_LOG_METRICS` flag.