Add ability to set pool size from HTTP API
Factored out HTTP API and put it behind a Gorilla mux.
Changed API to be more machine-friendly, using JSON for serialization.
Changed how ProcessorPool reasons about its size, allowing proper
handling of processors that are gracefully shutting down.
Added ability to start a worker with pool size = 0, which waits
indefinitely for a processor to be started.