Don't unescape ACCOUNT_JSON variable
The GCE provider uses an ACCOUNT_JSON variable to fetch the credentials
for accessing the GCE API. This variable can either be a path to the
JSON or a string of the JSON. A similar implementation is used in
gcloud-cleanup.

In worker, the provider config is processed and all config variables are
'unescaped' as if it was a URL. This causes the JSON to become unusable,
mainly because the private key contains `+` symbols which get replaced
by spaces.

This solution simply checks if the key ends with ACCOUNT_JSON and skips
the 'unescape' step. This ensures backwards compatibility for any other
variables relying on this behavior. Ideally we should only run this step
when required, as in 'variable is a URL that needs unescaping'.